requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
passlib
python-jose[cryptography]
python-multipart
email-validator
bcrypt